pyproject.toml
[tool.poetry]
name = "ragbox-core"
version = "1.0.0"
description = "RAG-in-a-Box: Zero-Configuration Self-Building Agentic RAG System"
authors = ["Developer <amankumarpandeyin@gmail.com>"]
readme = "README.md"
packages = [{include = "ragbox"}]
exclude = ["tests/", "docs/", ".env", "*.db", "chroma_db/", "knowledge_graph/", "logs/"]
[tool.poetry.scripts]
ragbox = "ragbox.cli:app"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0"
aiofiles = "^23.0"
watchdog = "^3.0"
chromadb = "^0.4"
neo4j = "^5.0"
networkx = "^3.0"
scikit-learn = "^1.3"
pypdf = "*"
pdfplumber = "*"
camelot-py = "*"
tree-sitter = "*"
tree-sitter-python = "*"
paddleocr = "*"
sentence-transformers = "*"
rank-bm25 = "*"
numpy = "^1.24"
scipy = "^1.11"
loguru = "^0.7"
typer = "^0.9"
rich = "^13.0"
llama-cpp-python = "*"
openai = "*"
anthropic = "*"
groq = "*"
python-dotenv = "*"
python-pptx = "*"
tenacity = "*"
tiktoken = "*"
python-magic = "*"
leidenalg = "*"
python-louvain = "*"
igraph = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
mypy = "^1.5.0"
black = "^23.7.0"
ruff = "^0.0.285"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
